Big Changes.
In order to do the ml_edge_restriction properly, we must restrict the fluxes, not the edge state variables.

Now, mkflux returns both edge states AND fluxes (is is_conservative(n) is true).

In update, we pass in both edge states and fluxes.  And, if is_conservative, we difference the fluxes.  Otherwise, we still have edge states for non-conservative differencing.

Still haven't added the call to ml_edge_restriction.  That's coming soon.

This also leads to the open question of how to deal with "flux" synchronization for non-conservative differencing.  Right now, it isn't being done.